requirements.txt
cachetools==3.0.0
google-api-python-client==1.7.7
google-auth==1.6.2
google-auth-httplib2==0.0.3
httplib2==0.12.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pyrotools==1.0.0
python-dateutil==2.7.5
rsa==4.0
six==1.12.0
uritemplate==3.0.0
